feat: implement inability adding to favorites without auth

+++ src/components/MovieCard/MovieCard.tsx
@@ -1,10 +1,10 @@
-import { memo, useEffect } from "react";
-import { generatePath } from "react-router-dom";
+import { memo } from "react";
+import { generatePath, useNavigate } from "react-router-dom";
 
 import { MovieLink, Poster } from "components";
 import { ROUTE } from "router";
 import { Movie } from "types";
-import { UseAppDispatch } from "store/hooks";
+import { UseAppDispatch, useAppSelector } from "store/hooks";
 import { addToFavorites, deleteFromFavorites, setFavorites, toggleFavorite } from "store/features";
 
 import {
@@ -21,19 +21,21 @@ interface MovieCardProps {
 
 export const MovieCard = memo(
   ({ movie: { poster, title, type, year, imdbID, isFavorite }, movie }: MovieCardProps) => {
+    const { isAuth } = useAppSelector((state) => state.user);
     const dispatch = UseAppDispatch();
-
+    const navigate = useNavigate();
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-      if (e.target.checked) {
+      if (isAuth && e.target.checked) {
         dispatch(addToFavorites(movie));
         dispatch(setFavorites(movie));
         dispatch(toggleFavorite({ imdbID, isFavorite: true }));
       }
-      if (!e.target.checked) {
+      if (isAuth && !e.target.checked) {
         dispatch(toggleFavorite({ imdbID, isFavorite: false }));
         dispatch(deleteFromFavorites(imdbID));
         dispatch(setFavorites(movie));
       }
+      !isAuth && navigate(ROUTE.SIGN_IN);
     };
 
     return (
@@ -42,12 +44,21 @@ export const MovieCard = memo(
           <Poster src={poster} alt={title} />
           {title}
         </MovieLink>
-        <StyledCheckbox
-          icon={<BorderIcon />}
-          checkedIcon={<CheckedIcon />}
-          onChange={handleChange}
-          checked={isFavorite}
-        />
+        {isAuth ? (
+          <StyledCheckbox
+            icon={<BorderIcon />}
+            checkedIcon={<CheckedIcon />}
+            onChange={handleChange}
+            checked={isFavorite}
+          />
+        ) : (
+          <StyledCheckbox
+            icon={<BorderIcon />}
+            checkedIcon={<BorderIcon />}
+            onChange={handleChange}
+            checked={isFavorite}
+          />
+        )}
         <StyledMovieYear>{year}</StyledMovieYear>
       </StyledMovieCard>
     );
